+++ package.json
@@ -6,7 +6,7 @@
   "types": "dist/server.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "rm -rf dist && tsc",
+    "build": "tsc",
     "serve": "ts-node src/server.ts",
     "dev": "ts-node-dev src/server.ts",
     "postinstall": "node ./scripts/postinstall"
@@ -16,9 +16,7 @@
   "dependencies": {
     "koa": "^2.10.0",
     "koa-router": "^7.4.0",
-    "lodash": "^4.17.15",
-    "rtsp-live555": "^1.8.7",
-    "validator": "^11.1.0"
+    "rtsp-live555": "^1.8.7"
   },
   "devDependencies": {
     "@types/koa": "^2.0.51",
