+++ src/server.ts
@@ -1,14 +1,53 @@
 import * as Koa from 'koa'
 import * as Router from 'koa-router';
+import {PassThrough} from 'stream'
+import liveServer from './lib/index';
 import config from './config'
-import api from './api'
 
 const app = new Koa();
+const router = new Router();
 
-const route = new Router();
-route.use('/api', api.routes(), api.allowedMethods());
+router.get('/*', async (ctx) => {
+    const url = ctx.request.url.slice(1);
+    if (url) {
+        ctx.set({
+            'Access-Control-Allow-Origin': '*',
+            'Connection': 'Keep-Alive',
+            'Content-Type': 'video/x-flv'
+        });
+        // ctx.status = 200;
+        let stream = new liveServer(url);
+        const _stream = new PassThrough();
+        stream.on('start', () => {
+            console.log(url + ' started');
+        });
+        stream.on('stop', () => {
+            console.log(url + ' stopped');
+        });
+        stream.on('data', (data) => {
+            let _success: boolean = _stream.write(data);
+        });
+        _stream.on('close', () => {
+            console.log('passThrough-close');
+            _stream.end();
+            stream.stop();
+            ctx.res.end()
+        });
+        // .on('finish', () => {
+        //     console.log('passThrough-finish');
+        //     _stream.end();
+        //     stream.stop();
+        //     ctx.res.end()
+        // });
 
-app.use(route.routes()).use(route.allowedMethods());
+        ctx.body = _stream
+    } else {
+        ctx.status = 204;
+        ctx.body = null;
+    }
+});
+
+app.use(router.routes());
 
 app.listen(config.port);
